Cosolidate api files
